Update GitHub API usage
We received a deprecation notice about our usage of the GitHub API and
how we were passing the authentication token.

The only reason we were manually calling the GitHub API is because at
one point in the past, the GitHub API Ruby gem was unable to retrieve
what we needed from the API. But those days are gone!

The current version of the GitHub API gem now does what we want, so
let's update our code to use that!